Move zipper from slc to sendletter package
