+++ src/main/java/uk/gov/hmcts/reform/sendletter/services/zip/Zipper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.slc.services.steps.zip;
+package uk.gov.hmcts.reform.sendletter.services.zip;
 
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.slc.services.steps.getpdf.PdfDoc;
