+++ src/test/java/uk/gov/hmcts/reform/sendletter/services/zip/ZipFileNameHelperTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.slc.services.steps.zip;
+package uk.gov.hmcts.reform.sendletter.services.zip;
 
 import org.junit.Test;
 import uk.gov.hmcts.reform.sendletter.entity.Letter;
